@@ -1,159 +1,249 @@
+import 'dart:async';
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_blue_plus/flutter_blue_plus.dart';
 
 import 'device_screen.dart';
 
 void main() {
-  runApp(MyApp());
+  runApp(const MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  final title = 'BLE Set Notification';
+  const MyApp({super.key});
+
+  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: title,
-      home: MyHomePage(title: title),
+      title: 'Flutter Demo',
+      theme: ThemeData(
+        // This is the theme of your application.
+        //
+        // TRY THIS: Try running your application with "flutter run". You'll see
+        // the application has a blue toolbar. Then, without quitting the app,
+        // try changing the seedColor in the colorScheme below to Colors.green
+        // and then invoke "hot reload" (save your changes or press the "hot
+        // reload" button in a Flutter-supported IDE, or press "r" if you used
+        // the command line to start the app).
+        //
+        // Notice that the counter didn't reset back to zero; the application
+        // state is not lost during the reload. To reset the state, use hot
+        // restart instead.
+        //
+        // This works for code too, not just values: Most code changes can be
+        // tested with just a hot reload.
+        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
+        useMaterial3: true,
+      ),
+      home: const DeviceSelectionScreen(),
     );
   }
 }
 
-class MyHomePage extends StatefulWidget {
-  MyHomePage({Key? key, required this.title}) : super(key: key);
-  final String title;
+class DeviceSelectionScreen extends StatefulWidget {
+  const DeviceSelectionScreen({super.key});
 
   @override
-  _MyHomePageState createState() => _MyHomePageState();
+  _DeviceSelectionScreenState createState() => _DeviceSelectionScreenState();
 }
 
-class _MyHomePageState extends State<MyHomePage> {
+class _DeviceSelectionScreenState extends State<DeviceSelectionScreen> {
   FlutterBluePlus flutterBlue = FlutterBluePlus.instance;
-  List<ScanResult> scanResultList = [];
-  bool _isScanning = false;
+  List<ScanResult> scanResults = [];
+  bool isScanning = false;
+
+
 
   @override
-  initState() {
+  void initState() {
     super.initState();
-    // 블루투스 초기화
-    initBle();
+    startScan();
   }
 
-  void initBle() {
-    // BLE 스캔 상태 얻기 위한 리스너
-    flutterBlue.isScanning.listen((isScanning) {
-      _isScanning = isScanning;
-      setState(() {});
+  void startScan() {
+    flutterBlue.startScan(timeout: const Duration(seconds: 4));
+
+    // Abonnieren Sie die Scan-Results und aktualisieren Sie die Liste
+    flutterBlue.scanResults.listen((List<ScanResult> results) {
+      setState(() {
+        scanResults = results;
+      });
+    });
+
+    setState(() {
+      isScanning = true;
     });
   }
 
-  /*
-  스캔 시작/정지 함수
-  */
-  scan() async {
-    if (!_isScanning) {
-      // 스캔 중이 아니라면
-      // 기존에 스캔된 리스트 삭제
-      scanResultList.clear();
-      // 스캔 시작, 제한 시간 4초
-      flutterBlue.startScan(timeout: Duration(seconds: 4));
-      // 스캔 결과 리스너
-      flutterBlue.scanResults.listen((results) {
-        scanResultList = results;
-        // UI 갱신
-        setState(() {});
-      });
-    } else {
-      // 스캔 중이라면 스캔 정지
-      flutterBlue.stopScan();
-    }
+  void stopScan() {
+    flutterBlue.stopScan();
+    setState(() {
+      isScanning = false;
+    });
   }
 
-  /*
-   여기서부터는 장치별 출력용 함수들
-  */
-  /*  장치의 신호값 위젯  */
-  Widget deviceSignal(ScanResult r) {
-    return Text(r.rssi.toString());
-  }
 
-  /* 장치의 MAC 주소 위젯  */
-  Widget deviceMacAddress(ScanResult r) {
-    return Text(r.device.id.id);
-  }
 
-  /* 장치의 명 위젯  */
-  Widget deviceName(ScanResult r) {
-    String name = '';
+  void connectToDevice(BluetoothDevice device, context) {
+    StreamSubscription<BluetoothDeviceState> subscription;
+    device.connect();
 
-    if (r.device.name.isNotEmpty) {
-      // device.name에 값이 있다면
-      name = r.device.name;
-    } else if (r.advertisementData.localName.isNotEmpty) {
-      // advertisementData.localName에 값이 있다면
-      name = r.advertisementData.localName;
-    } else {
-      // 둘다 없다면 이름 알 수 없음...
-      name = 'N/A';
-    }
-    return Text(name);
-  }
+    subscription = device.state.listen((state) async {
+      if (state == BluetoothDeviceState.connected) {
+        print('Verbindung hergestellt mit ${device.name}');
 
-  /* BLE 아이콘 위젯 */
-  Widget leading(ScanResult r) {
-    return CircleAvatar(
-      child: Icon(
-        Icons.bluetooth,
-        color: Colors.white,
-      ),
-      backgroundColor: Colors.cyan,
-    );
-  }
+        Navigator.pushReplacement(
+          context,
+          MaterialPageRoute(builder: (context) => MyCustomWidget( device: device)),
+        );
 
-  /* 장치 아이템을 탭 했을때 호출 되는 함수 */
-  void onTap(ScanResult r) {
-    // 단순히 이름만 출력
-    print('${r.device.name}');
-    Navigator.push(
-      context,
-      MaterialPageRoute(builder: (context) => DeviceScreen(device: r.device)),
-    );
-  }
+        // Führen Sie hier Ihre erforderlichen Aktionen mit der Verbindung durch
+        // Beispiel: Daten senden/empfangen, Charakteristiken lesen/schreiben, usw.
 
-  /* 장치 아이템 위젯 */
-  Widget listItem(ScanResult r) {
-    return ListTile(
-      onTap: () => onTap(r),
-      leading: leading(r),
-      title: deviceName(r),
-      subtitle: deviceMacAddress(r),
-      trailing: deviceSignal(r),
-    );
+        // Trennen Sie die Verbindung, wenn sie nicht mehr benötigt wird
+        // await device.disconnect();
+
+      } else if (state == BluetoothDeviceState.disconnected) {
+        print('Fehler beim Verbinden mit ${device.name}');
+      }
+    });
   }
 
   @override
   Widget build(BuildContext context) {
+    print('started');
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: const Text('Device Selection'),
       ),
-      body: Center(
-        /* 장치 리스트 출력 */
-        child: ListView.separated(
-          itemCount: scanResultList.length,
-          itemBuilder: (context, index) {
-            return listItem(scanResultList[index]);
-          },
-          separatorBuilder: (BuildContext context, int index) {
-            return Divider();
-          },
-        ),
-      ),
-      /* 장치 검색 or 검색 중지  */
-      floatingActionButton: FloatingActionButton(
-        onPressed: scan,
-        // 스캔 중이라면 stop 아이콘을, 정지상태라면 search 아이콘으로 표시
-        child: Icon(_isScanning ? Icons.stop : Icons.search),
+      body: Column(
+        children: [
+          ElevatedButton(
+            onPressed: isScanning ? stopScan : startScan,
+            child: Text(isScanning ? 'Stop Scan' : 'Start Scan'),
+          ),
+          Expanded(
+            child: ListView.builder(
+              itemCount: scanResults.length,
+              itemBuilder: (context, index) {
+                ScanResult scanResult = scanResults[index];
+                BluetoothDevice device = scanResult.device;
+                return ListTile(
+                  title: Text(device.name),
+                  subtitle: Text(device.id.toString()),
+                  onTap: () {
+                    connectToDevice(device, context);
+                  },
+                );
+              },
+            ),
+          ),
+        ],
       ),
     );
   }
+}
+
+class MyCustomWidget extends StatefulWidget {
+  final BluetoothDevice device;
+
+  const MyCustomWidget({super.key, required this.device});
+  @override
+  _MyCustomWidgetState createState() => _MyCustomWidgetState();
+}
+
+class _MyCustomWidgetState extends State<MyCustomWidget> {
+  late BluetoothCharacteristic characteristic;
+  late StreamSubscription<List<int>> subscription;
+  String access = 'Warte auf Antwort...';
+
+  @override
+  Widget build(BuildContext context) {
+    BluetoothDevice device = widget.device;
+
+    return Scaffold(
+      appBar: AppBar(
+        title: Text(device.name),
+      ),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: [
+            const Text('Connected to:'),
+            Text('Name: ${device.name}'),
+            Text('Adresse: ${device.id.toString()}'),
+            const SizedBox(height: 16),
+            Text(access),
+            const SizedBox(height: 16),
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: [
+                ElevatedButton(
+                  onPressed: () {
+                    sendDataToDoor(device);
+                  },
+                  child: const Text('Get Access'),
+
+                ),
+                ElevatedButton(
+                  onPressed: () {
+                    try {
+                      subscription.cancel();
+                    }catch(e){
+                      // Hhh
+                    }
+                    device.disconnect();
+                    Navigator.pushReplacement(
+                      context,
+                      MaterialPageRoute(builder: (context) => const DeviceSelectionScreen()),
+                    );
+                  },
+                  child: const Text('Disconnect'),
+
+                ),
+              ],
+            )
+          ],
+        ),
+      ),
+    );
+  }
+
+
+
+  void sendDataToDoor(BluetoothDevice device) async {
+    List<BluetoothService> services = await device.discoverServices();
+
+    // Suchen Sie die gewünschte Charakteristik in den gefundenen Diensten
+    for (BluetoothService service in services) {
+      for (BluetoothCharacteristic c in service.characteristics) {
+        if (c.uuid.toString() == "00000002-0000-1000-8000-00805f9b34fb") {
+          characteristic = c;
+        }
+      }
+
+    }
+    String request = "Hello, world!"; // Der UTF-8-Request, den du senden möchtest
+    List<int> data = utf8.encode(request);
+    print(characteristic.uuid.toString()); // UTF-8-Request in Bytes umwandeln
+    try {
+      await characteristic.setNotifyValue(true);
+    }catch(e){
+      //hhekjkk
+    }
+
+    subscription = characteristic.value.listen((value) {
+      String response = utf8.decode(value);
+      setState(() {
+        access = response;
+      });
+      print("Antwort erhalten: $response");
+      // Hier kannst du die empfangenen Daten weiterverarbeiten
+    });
+    await characteristic.write(data);
+
+  }
+
 }
